ASSETS-11385 add fingerprint logic to aio-cli-plugin-certificate
