+++ package.json
@@ -15,8 +15,7 @@
     "@oclif/errors": "^1.1.2",
     "@oclif/plugin-help": "^2.2.3",
     "cli-ux": "^5.4.5",
-    "js-yaml": "^3.13.1",
-    "node-forge": "^1.3.0"
+    "js-yaml": "^3.13.1"
   },
   "devDependencies": {
     "@adobe/eslint-config-aio-lib-config": "^1.4.0",
