+++ src/commands/console/publickey/upload.js
@@ -11,14 +11,12 @@ governing permissions and limitations under the License.
 */
 const { flags } = require('@oclif/command')
 const ConsoleCommand = require('../index')
+const IndexCommand = require('./index')
 const fs = require('fs')
 const { cli } = require('cli-ux')
 const { CONFIG_KEYS } = require('../../../config')
 const LibConsoleCLI = require('@adobe/aio-cli-lib-console')
 const aioConsoleLogger = require('@adobe/aio-lib-core-logging')('@adobe/aio-cli-plugin-console:workspace:bind', { provider: 'debug' })
-const forge = require('node-forge')
-const pki = forge.pki
-const asn1 = forge.asn1
 
 class UploadAndBindCommand extends ConsoleCommand {
   async run () {
@@ -63,7 +61,7 @@ class UploadAndBindCommand extends ConsoleCommand {
       const workspace = project.workspace
 
       const certPemString = fs.readFileSync(publicKey)
-      const fingerprint = this.computeFingerprint(certPemString)
+      const { certificateFingerprint: fingerprint } = await this.consoleCLI.getCertificateFingerprint(certPemString)
 
       const existing = await this.consoleCLI.getBindingsForWorkspace(orgId, project, workspace)
 
@@ -87,7 +85,7 @@ class UploadAndBindCommand extends ConsoleCommand {
       } else if (flags.yml) {
         this.printYaml(bindings)
       } else {
-        this.printResults(bindings)
+        IndexCommand.printBindings(bindings)
       }
       return bindings
     } catch (err) {
@@ -96,24 +94,6 @@ class UploadAndBindCommand extends ConsoleCommand {
     } finally {
       LibConsoleCLI.cleanStdOut()
     }
-  }
-
-  /**
-   * Computes the SHA-1 digest of the entire DER-encoded x.509 certificate
-   * contained in the provided PEM-encoded string, which gives the same result as
-   * using "openssl x509 -fingerprint", except without delimeters and in
-   * all lowercase.
-   *
-   * @param {string} certPemString PEM-encoded sting containing x509 certificate
-   * @returns {string} x509 fingerprint
-   */
-  computeFingerprint (certPemString) {
-    const cert = pki.certificateFromPem(certPemString, false, false)
-    const bytes = asn1.toDer(pki.certificateToAsn1(cert)).getBytes()
-    const md = forge.md.sha1.create()
-    md.start()
-    md.update(bytes)
-    return md.digest().toHex()
   }
 
   printResults (bindings) {
